remove ignored specs, move retro and timeline to new parent
