+++ webapp/src/components/backstage/playbook_runs/playbook_run/retrospective/report.tsx
@@ -7,8 +7,7 @@ import styled from 'styled-components';
 import {useIntl} from 'react-intl';
 
 import {PlaybookRun} from 'src/types/playbook_run';
-import ReportTextArea
-    from 'src/components/backstage/playbook_runs/playbook_run_backstage/retrospective/report_text_area';
+import ReportTextArea from 'src/components/backstage/playbook_runs/playbook_run/retrospective/report_text_area';
 
 interface ReportProps {
     playbookRun: PlaybookRun;
