+++ webapp/src/components/backstage/playbook_runs/playbook_run/rhs_timeline.tsx
@@ -12,7 +12,7 @@ import {useIntl} from 'react-intl';
 import Scrollbars from 'react-custom-scrollbars';
 
 import {renderThumbVertical, renderTrackHorizontal, renderView} from '../../../rhs/rhs_shared';
-import TimelineEventItem from 'src/components/backstage/playbook_runs/playbook_run_backstage/retrospective/timeline_event_item';
+import TimelineEventItem from 'src/components/backstage/playbook_runs/playbook_run/retrospective/timeline_event_item';
 import {PlaybookRun} from 'src/types/playbook_run';
 import {clientRemoveTimelineEvent} from 'src/client';
 import MultiCheckbox from 'src/components/multi_checkbox';
